+++ MonoGame.Framework/Input/Joystick.Web.cs
@@ -45,7 +45,7 @@ namespace Microsoft.Xna.Framework.Input
         private static JoystickState PlatformGetState(int index)
         {
             var connected = false;
-            var axes = new float[0];
+            var axes = new int[0];
             var buttons = new ButtonState[0];
 
             var navigator = Builtins.Global["navigator"];
@@ -58,7 +58,7 @@ namespace Microsoft.Xna.Framework.Input
                     connected = true;
 
                     var axescount = gamepads[index].axes.length;
-                    axes = new float[gamepads[index].axes.length];
+                    axes = new int[gamepads[index].axes.length];
 
                     for (int i = 0; i < axescount; i++)
                         axes[i] = gamepads[index].axes[i];
